requirements.txt
colorama==0.4.4
datetimerange==1.2.0
fastapi==0.68.2
gunicorn==20.1.0
loguru==0.5.3
mariadb==1.0.8
py==1.10.0
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.1
python-levenshtein==0.12.2
pyzmq==22.3.0
sqlalchemy==1.4.26
starlette==0.14.2
typepy==1.3.0
typing-extensions==3.10.0.2
uvicorn==0.15.0